Order endpoints results done.

+++ WebbLabb2.DataAccess/Repositories/OrderRepository.cs
@@ -1,5 +1,8 @@
-﻿using MongoDB.Driver;
+﻿using Microsoft.EntityFrameworkCore;
+using MongoDB.Bson;
+using MongoDB.Driver;
 using WebbLabb2.DataAccess.Models;
+using WebbLabb2.DataAccess.Sql.Contexts;
 using WebbLabb2.Shared.DTOs;
 
 namespace WebbLabb2.DataAccess.Repositories;
@@ -7,9 +10,12 @@ namespace WebbLabb2.DataAccess.Repositories;
 public class OrderRepository
 {
     private readonly IMongoCollection<OrderModel> _orderModelCollection;
+    private readonly CustomerDbContext _customerDbContext;
 
-    public OrderRepository()
+    public OrderRepository(CustomerDbContext customerDbContext)
     {
+        _customerDbContext = customerDbContext;
+
         var host = "localhost";
         var databaseName = "OrderDb";
         var port = 27017;
@@ -21,24 +27,35 @@ public class OrderRepository
                 new MongoCollectionSettings() { AssignIdOnInsert = true });
     }
 
-    public async Task CreateOrder(OrderDto dto)
+    public async Task<bool> CreateOrder(OrderDto dto)
     {
+        var validEmail = await _customerDbContext.CustomerModel.AnyAsync(c => c.Email == dto.CustomerEmail);
+        if (validEmail == false) return false;
+
         await _orderModelCollection.InsertOneAsync(ConvertToModel(dto));
+        return true;
     }
 
-    public async Task UpdateOrder(string id, OrderDto dto)
+    public async Task<bool> UpdateOrder(string id, OrderDto dto)
     {
         var filter = Builders<OrderModel>.Filter.Eq("OrderId", id);
         var update = Builders<OrderModel>.Update
             .Set(o => o.ProductList, dto.ProductList);
 
+        var orderIdExists = await _orderModelCollection.Find(filter).AnyAsync();
+        if (!orderIdExists) return false;
+
         await _orderModelCollection.UpdateOneAsync(filter, update);
+        return true;
     }
 
     public async Task<OrderDto[]> GetOrder(string id)
     {
         var order = await _orderModelCollection.FindAsync(o => o.OrderId.Equals(id));
-        return order.ToList().Select(ConvertToDto).ToArray();
+        
+        var orderIdExists = await _orderModelCollection.FindAsync(o => o.OrderId.Equals(id)).Result.AnyAsync();
+        
+        return !orderIdExists ? null : order.ToList().Select(ConvertToDto).ToArray();
     }
 
     public async Task<OrderDto[]> GetAllOrders()
@@ -48,9 +65,13 @@ public class OrderRepository
         return getAllOrders.ToList().Select(ConvertToDto).ToArray();
     }
 
-    public async Task RemoveOrder(string id)
+    public async Task<bool> RemoveOrder(string id)
     {
+        var orderIdExists = await _orderModelCollection.FindAsync(o => o.OrderId.Equals(id)).Result.AnyAsync();
+        if (orderIdExists == false) return false;
+
         await _orderModelCollection.DeleteOneAsync(p => p.OrderId.Equals(id));
+        return true;
     }
 
     private OrderModel ConvertToModel(OrderDto dto)
